requirements.txt
django==2.2.10
django-admin-interface==0.12.0
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
Faker==4.0.1
Pillow==6.0.0
psycopg2-binary==2.8.1
pytest==5.0.1
pytest-django==3.5.1
dj-database-url==0.5.0
gunicorn==19.9.0
django-storages==1.7.1
boto3==1.9.144
sparkpost==1.3.6
